moved middleware to separate file
