+++ server.js
@@ -1,9 +1,9 @@
 const express = require('express');
 const mongoose = require('mongoose');
 
-const PORT = process.env.PORT || 5000;
+// const PORT = process.env.PORT || 5000;
 
-// const PORT = 5000;
+const PORT = 5000;
 
 // ========== MIDDLEWARE ============== //
